soc: qcom: smem: Switch partitions to xarray
The partitions array is statically sized and uses the remote host ID as
its index. There are valid remote host IDs which greatly exceeds the
current defined size. Instead of increasing the size of the array and
wasting memory for large remote host IDs, switch over to using xarray.
This provides partitions xarray to grow dynamically and save memory.

Signed-off-by: Tony Truong <quic_truong@quicinc.com>
Signed-off-by: Shawn Guo <shengchao.guo@oss.qualcomm.com>